fix(aapd-1676): moved withdrawn appeals filter from common to business rules to address pr comment
