+++ packages/common/src/view-model-maps/index.js
@@ -6,5 +6,5 @@ module.exports = {
 	...require('./rows'),
 	...require('../lib/format-questionnaire-details'),
 	...require('./events'),
-	...require('../lib/filter-withdrawn-appeal')
+	...require('@pins/business-rules/src/lib/filter-withdrawn-appeal')
 };
